pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Network_devices_management_system"
version = "0.1.0"
description = "基于 Fastapi 构建的网络设备管理系统"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "lijianqiao", email = "lijianqiao2906@live.com" }]
dependencies = [
    "aerich>=0.9.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "redis[hiredis]>=6.2.0",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.34.3",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py313"
exclude = [".git", "__pycache__", ".venv", ".vscode"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle 错误
    "W",  # PycodeStyle 警告
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8 推导式
    "UP", # pyupgrade
]
ignore = [
    "E501", # 行太长，由 black 处理
    "B008", # 不在参数默认值中执行函数调用
    "W191", # 缩进包含制表符
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

# 项目级将 uv 下载源换成国内源 - 清华大学
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[dependency-groups]
dev = ["pytest>=8.4.0", "pytest-asyncio>=1.0.0"]

[tool.aerich]
tortoise_orm = "app.db.connection.TORTOISE_ORM"
location = "./migrations"
src_folder = "." 
